+++ core/src/AzureMcp.Core/AssemblyInfo.cs
@@ -6,3 +6,4 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("AzureMcp.Core.UnitTests")]
 [assembly: InternalsVisibleTo("AzureMcp.Core.LiveTests")]
 [assembly: InternalsVisibleTo("AzureMcp.Tests")]
+[assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]